Add instructions for custom server
This took me 2-3 hours to fully figure out.  I was trying to figure out how to serve the React files from a custom Hapi server, and there didn't seem to be any easy way to merge the two TS files.

These instructions make it an extremely simple process.  Since I would have loved to have this information earlier, and I expect many users will want to use their own server for serving the files, I think this will be a useful change!

This is the preferred method from this article: https://www.fullstackreact.com/articles/using-create-react-app-with-a-server/